+++ vortex-serde/src/layouts/read/filtering.rs
@@ -2,16 +2,17 @@ use std::any::Any;
 use std::fmt::Debug;
 use std::sync::Arc;
 
+use arrow_buffer::BooleanBuffer;
 use vortex::aliases::hash_set::HashSet;
-use vortex::array::ConstantArray;
+use vortex::array::{BoolArray, ConstantArray};
 use vortex::compute::and;
 use vortex::stats::ArrayStatistics;
+use vortex::validity::Validity;
 use vortex::{Array, IntoArray, IntoArrayVariant};
 use vortex_dtype::field::Field;
 use vortex_error::{VortexExpect, VortexResult};
 use vortex_expr::{split_conjunction, unbox_any, VortexExpr};
 
-use crate::layouts::null_as_false;
 use crate::layouts::read::filter_project::filter_project;
 
 #[derive(Debug, Clone)]
@@ -80,5 +81,41 @@ impl PartialEq<dyn Any> for RowFilter {
             .downcast_ref::<Self>()
             .map(|x| x == other)
             .unwrap_or(false)
+    }
+}
+
+pub fn null_as_false(array: BoolArray) -> VortexResult<Array> {
+    Ok(match array.validity() {
+        Validity::NonNullable => array.into_array(),
+        Validity::AllValid => {
+            BoolArray::try_new(array.boolean_buffer(), Validity::NonNullable)?.into_array()
+        }
+        Validity::AllInvalid => BoolArray::from(BooleanBuffer::new_unset(array.len())).into_array(),
+        Validity::Array(v) => {
+            let bool_buffer = &array.boolean_buffer() & &v.into_bool()?.boolean_buffer();
+            BoolArray::from(bool_buffer).into_array()
+        }
+    })
+}
+
+#[cfg(test)]
+mod tests {
+    use vortex::array::BoolArray;
+    use vortex::validity::Validity;
+    use vortex::IntoArrayVariant;
+
+    use super::*;
+
+    #[test]
+    fn coerces_nulls() {
+        let bool_array = BoolArray::from_vec(
+            vec![true, true, false, false],
+            Validity::Array(BoolArray::from(vec![true, false, true, false]).into()),
+        );
+        let non_null_array = null_as_false(bool_array).unwrap().into_bool().unwrap();
+        assert_eq!(
+            non_null_array.boolean_buffer().iter().collect::<Vec<_>>(),
+            vec![true, false, false, false]
+        );
     }
 }
